feat(aws-smithy-types): Decouple http-body-1-x from http 0.2.x
The http-body-1-x feature no longer enables the http 0.2.x or
http-body 0.4.x crates. The legacy 0.4.x adapter code (the
Http1toHttp04 body adapter and the 0.2.x header conversions) is now
gated behind http-body-0-4-x, and rt-tokio continues to enable the
0.4.x stack it needs for file-based bodies. The http-body 1.x path is
now free of the http 0.2.x crate.

Together with aws-smithy-runtime-api gating its http 0.2.x dependency
behind http-02x, this removes http 0.2.x from the default dependency
tree of the runtime crates, addressing the unpatched http 0.2.x
advisories.

sim: https://github.com/smithy-lang/smithy-rs/issues/4805
